Refactor Cask model
