+++ Applite/Model/Cask Models/BrewAnalytics.swift
